+++ src/Services/Nokia/NokiaService.php
@@ -90,6 +90,10 @@ class NokiaService
 
     public function disconnect(): void
     {
+        if (self::$telnetConn === null && self::$tl1Conn === null) {
+            throw new Exception('No connection established.');
+        }
+
         if (self::$telnetConn !== null) {
             self::$telnetConn->destroy();
             self::$telnetConn = null;
@@ -98,10 +102,6 @@ class NokiaService
         if (self::$tl1Conn !== null) {
             self::$tl1Conn->destroy();
             self::$tl1Conn = null;
-        }
-
-        if (self::$telnetConn === null && self::$tl1Conn === null) {
-            throw new Exception('No connection established.');
         }
     }
